pyproject.toml
[tool.poetry]
name = "tz-second-part"
version = "0.1.0"
description = ""
authors = ["emotzart"]
readme = "README.md"
packages = [{include = "tz_second_part"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
